requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4.post0
    # via discord-py
async-timeout==3.0.1
    # via aiohttp
attrs==21.4.0
    # via aiohttp
chardet==4.0.0
    # via aiohttp
discord-py==1.7.3
    # via -r requirements.in
idna==3.3
    # via yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
python-dotenv==0.20.0
    # via -r requirements.in
typing-extensions==4.2.0
    # via aiohttp
yarl==1.7.2
    # via aiohttp
